completed

+++ admin/package.php
@@ -102,12 +102,15 @@
                                                 <a href="package_edit.php?package_id=<?php echo $row['package_id']; ?>"
                                                     class="btn btn-darkblue">
                                                     <i class="fa-solid fa-pen-to-square"></i> </a>
-                                                <?php if ($row['package_status']=="0"): ?>
-                                                <button type="button"
-                                                    onclick="deleteData(<?php echo $row['package_id']; ?>,'package', 'package_id')"
-                                                    class="btn btn-darkblue"> <i class="fa-solid fa-trash"></i>
-                                                </button>
-                                                <?php endif ?>
+
+                                                <?php if (isset($_SESSION['admin']) && $_SESSION['admin'] == 'admin'): ?>
+                                                    <?php if ($row['package_status']=="0"): ?>
+                                                    <button type="button"
+                                                        onclick="deleteData(<?php echo $row['package_id']; ?>,'package', 'package_id')"
+                                                        class="btn btn-darkblue"> <i class="fa-solid fa-trash"></i>
+                                                    </button>
+                                                    <?php endif; ?>
+                                                <?php endif; ?>
 
                                             </td>
                                             <?php } ?>
